# swagger_server/models/light.py
# coding: utf-8

from __future__ import absolute_import
# from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model

from sqlalchemy import Table, Column, ForeignKey, String, Integer, Text
from sqlalchemy.orm import relationship, backref
from swagger_server.models.base import Base


class Light(Base):
    __tablename__ = 'lights'

    id = Column(Integer, primary_key=True)
    name = Column(String(128), primary_key=True)
    status = Column(Integer, primary_key=True)

# class Light(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    # def __init__(self, id: int=None, name: str=None, status: bool=False):
    #     """
    #     Light - a model defined in Swagger
    #
    #     :param id: The id of this Light.
    #     :type id: int
    #     :param name: The name of this Light.
    #     :type name: str
    #     :param status: The status of this Light.
    #     :type status: bool
    #     """
    #     self.swagger_types = {
    #         'id': int,
    #         'name': str,
    #         'status': bool
    #     }
    #
    #     self.attribute_map = {
    #         'id': 'id',
    #         'name': 'name',
    #         'status': 'status'
    #     }
    #
    #     self._id = id
    #     self._name = name
    #     self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'Light':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Light of this Light.
        :rtype: Light
        """
        return deserialize_model(dikt, cls)

    # @property
    # def id(self) -> int:
    #     """
    #     Gets the id of this Light.
    #
    #     :return: The id of this Light.
    #     :rtype: int
    #     """
    #     return self._id
    #
    # @id.setter
    # def id(self, id: int):
    #     """
    #     Sets the id of this Light.
    #
    #     :param id: The id of this Light.
    #     :type id: int
    #     """
    #
    #     self._id = id
    #
    # @property
    # def name(self) -> str:
    #     """
    #     Gets the name of this Light.
    #
    #     :return: The name of this Light.
    #     :rtype: str
    #     """
    #     return self._name
    #
    # @name.setter
    # def name(self, name: str):
    #     """
    #     Sets the name of this Light.
    #
    #     :param name: The name of this Light.
    #     :type name: str
    #     """
    #
    #     self._name = name
    #
    # @property
    # def status(self) -> bool:
    #     """
    #     Gets the status of this Light.
    #
    #     :return: The status of this Light.
    #     :rtype: bool
    #     """
    #     return self._status
    #
    # @status.setter
    # def status(self, status: bool):
    #     """
    #     Sets the status of this Light.
    #
    #     :param status: The status of this Light.
    #     :type status: bool
    #     """
    #
    #     self._status = status